refactor(routes): reorganize route files to folder-based structure
- Convert flat route files (e.g., [id].get.ts) to folder structure (e.g., [id]/index.get.ts)
- One endpoint per folder pattern for better code organization
- Update all affected routes: issues, prs, ca, cas, webhooks, health, config
- Update test files to match new folder structure
- All 355 tests passing
