+++ packages/hr-agent/src/routes/health/index.get.ts
@@ -1,5 +1,5 @@
 import type { Request, Response } from 'express';
-import Result from '../utils/Result.js';
+import Result from '../../utils/Result.js';
 
 export default function healthRoute(_req: Request, res: Response): void {
   res.json(new Result({ status: 'ok' }));
