+++ packages/hr-agent/src/routes/v1/config/index.get.ts
@@ -1,6 +1,6 @@
 import type { Request, Response } from 'express';
-import Result from '../../utils/Result.js';
-import { TASK_CONFIG } from '../../config/taskConfig.js';
+import Result from '../../../utils/Result.js';
+import { TASK_CONFIG } from '../../../config/taskConfig.js';
 
 export default function configRoute(_req: Request, res: Response): void {
   res.json(
