+++ packages/hr-agent/src/routes/v1/issues/[issueId]/latest-pr/index.get.ts
@@ -1,6 +1,6 @@
 import type { Request, Response } from 'express';
-import Result from '../../../../utils/Result.js';
-import { getPrismaClient, SOFT_DELETE_FLAG } from '../../../../utils/database.js';
+import Result from '../../../../../utils/Result.js';
+import { getPrismaClient, SOFT_DELETE_FLAG } from '../../../../../utils/database.js';
 
 const HTTP = {
   BAD_REQUEST: 400,
@@ -8,7 +8,7 @@ const HTTP = {
   NOT_FOUND: 404
 };
 
-export default async function getIssuePRsRoute(req: Request, res: Response): Promise<void> {
+export default async function getIssueLatestPRRoute(req: Request, res: Response): Promise<void> {
   const prisma = getPrismaClient();
   const { issueId } = req.params;
   const parsedIssueId = parseInt(issueId as string, 10);
@@ -19,7 +19,7 @@ export default async function getIssuePRsRoute(req: Request, res: Response): Pro
   }
 
   try {
-    const issuePrs = await prisma.issuePR.findMany({
+    const issuePr = await prisma.issuePR.findFirst({
       where: {
         issueId: parsedIssueId,
         issue: { deletedAt: SOFT_DELETE_FLAG },
@@ -31,11 +31,19 @@ export default async function getIssuePRsRoute(req: Request, res: Response): Pro
       orderBy: { createdAt: 'desc' }
     });
 
-    res.json(new Result(issuePrs.map((ip) => ip.pr)));
+    if (!issuePr) {
+      res.json(new Result().error(HTTP.NOT_FOUND, 'No PR found for this issue'));
+      return;
+    }
+
+    res.json(new Result(issuePr.pr));
   } catch (error) {
     const errorMessage = error instanceof Error ? error.message : 'Unknown error occurred';
     res.json(
-      new Result().error(HTTP.INTERNAL_SERVER_ERROR, `Failed to fetch issue PRs: ${errorMessage}`)
+      new Result().error(
+        HTTP.INTERNAL_SERVER_ERROR,
+        `Failed to fetch latest issue PR: ${errorMessage}`
+      )
     );
   }
 }
